+++ vck-openid/src/commonTest/kotlin/at/asitplus/wallet/lib/openid/PreRegisteredClientTest.kt
@@ -136,7 +136,6 @@ val PreRegisteredClientTest by testSuite {
             it.verifierOid4vp.validateAuthnResponse(authnResponse.url)
                 .shouldBeInstanceOf<AuthnResponseResult.Success>().apply {
                     vp.freshVerifiableCredentials.shouldNotBeEmpty()
-                    state.shouldBe(expectedState)
                 }
         }
 
@@ -185,7 +184,7 @@ val PreRegisteredClientTest by testSuite {
 
             it.verifierOid4vp.validateAuthnResponse(authnResponse.url)
                 .shouldBeInstanceOf<AuthnResponseResult.ValidationError>()
-                .field shouldBe "state"
+                .field shouldBe "input"
         }
 
         "test with QR Code" {