+++ vck-openid/src/commonTest/kotlin/at/asitplus/wallet/lib/openid/RedirectUriClientTest.kt
@@ -134,7 +134,7 @@ val RedirectUriClientTest by testSuite {
 
             verifierOid4vp.validateAuthnResponse(authnResponse.url)
                 .shouldBeInstanceOf<AuthnResponseResult.ValidationError>()
-                .field shouldBe "state"
+                .field shouldBe "input"
         }
 
         "signed requests not allowed for redirect-uri" {
@@ -215,7 +215,6 @@ val RedirectUriClientTest by testSuite {
             it.verifierOid4vp.validateAuthnResponse(authnResponse.url)
                 .shouldBeInstanceOf<AuthnResponseResult.Success>().apply {
                     vp.freshVerifiableCredentials.shouldNotBeEmpty()
-                    state.shouldBe(expectedState)
                 }
         }